+++ src/main/java/de/samply/email/EmailTemplatesFactory.java
@@ -7,8 +7,8 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.stereotype.Component;
 
-import java.io.File;
 import java.io.IOException;
+import java.nio.file.Path;
 
 @Slf4j
 @Component
@@ -16,9 +16,9 @@ public class EmailTemplatesFactory {
 
 
     @Bean
-    public EmailTemplates createEmailTemplates(@Value(ProjectManagerConst.EMAIL_TEMPLATES_CONFIG_PATH_SV) String templatesPath) {
+    public EmailTemplates createEmailTemplates(@Value(ProjectManagerConst.EMAIL_TEMPLATES_CONFIG_PATH_SV) Path templatesPath) {
         try {
-            return new ObjectMapper().readValue(new File(templatesPath), EmailTemplates.class);
+            return new ObjectMapper().readValue(templatesPath.toFile(), EmailTemplates.class);
         } catch (IOException e) {
             log.error("Configuration file for email templates not found: {}", templatesPath);
             throw new RuntimeException(e);