+++ src/main/java/de/samply/form/pdf/FormPdfConverter.java
@@ -4,21 +4,21 @@ import com.openhtmltopdf.pdfboxout.PdfRendererBuilder;
 import de.samply.app.ProjectManagerConst;
 import de.samply.pdf.PdfConverter;
 import de.samply.pdf.PdfConverterException;
+import de.samply.utils.directory.ExistingDirectory;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.OutputStream;
-import java.nio.file.Paths;
 
 @Component
 public class FormPdfConverter implements PdfConverter {
 
     private final String resourcesBaseUri;
 
-    public FormPdfConverter(@Value(ProjectManagerConst.FORM_RESOURCES_DIRECTORY_SV) String resourcesPath) {
-        this.resourcesBaseUri = Paths.get(resourcesPath).toUri().toString();
+    public FormPdfConverter(@Value(ProjectManagerConst.FORM_RESOURCES_DIRECTORY_SV) ExistingDirectory resourcesPath) {
+        this.resourcesBaseUri = resourcesPath.path().toUri().toString();
     }
 
     @Override