+++ src/main/java/de/samply/form/pdf/FormPdfGeneratorFactory.java
@@ -2,8 +2,8 @@ package de.samply.form.pdf;
 
 import de.samply.app.ProjectManagerConst;
 import de.samply.pdf.PdfGenerator;
-import de.samply.utils.DirectoryUtils;
 import de.samply.utils.FileExtension;
+import de.samply.utils.directory.ExistingDirectory;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import org.thymeleaf.TemplateEngine;
@@ -11,7 +11,6 @@ import org.thymeleaf.spring6.SpringTemplateEngine;
 import org.thymeleaf.templatemode.TemplateMode;
 import org.thymeleaf.templateresolver.FileTemplateResolver;
 
-import java.io.FileNotFoundException;
 import java.nio.charset.StandardCharsets;
 
 @Component
@@ -21,10 +20,9 @@ public class FormPdfGeneratorFactory {
     private final TemplateEngine templateEngine;
 
     public FormPdfGeneratorFactory(FormPdfConverter formPdfConverter,
-                                   @Value(ProjectManagerConst.FORM_RESOURCES_DIRECTORY_SV) String externalTemplateDirectory) throws FileNotFoundException {
+                                   @Value(ProjectManagerConst.FORM_RESOURCES_DIRECTORY_SV) ExistingDirectory externalTemplateDirectory) {
         this.formPdfConverter = formPdfConverter;
-        this.templateEngine = createTemplateEngine(
-                DirectoryUtils.fetchExternalTemplateDirectory(externalTemplateDirectory));
+        this.templateEngine = createTemplateEngine(externalTemplateDirectory.path().toAbsolutePath().toString());
     }
 
     public PdfGenerator createPdfGenerator() {