+++ src/main/java/de/samply/frontend/ActionExplanationsFactory.java
@@ -7,17 +7,17 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.stereotype.Component;
 
-import java.io.File;
 import java.io.IOException;
+import java.nio.file.Path;
 
 @Slf4j
 @Component
 public class ActionExplanationsFactory {
 
     @Bean
-    public ActionExplanations actionExplanations(@Value(ProjectManagerConst.ACTION_EXPLANATION_CONFIG_PATH_SV) String templatesPath) {
+    public ActionExplanations actionExplanations(@Value(ProjectManagerConst.ACTION_EXPLANATION_CONFIG_PATH_SV) Path templatesPath) {
         try {
-            return new ObjectMapper().readValue(new File(templatesPath), ActionExplanations.class);
+            return new ObjectMapper().readValue(templatesPath.toFile(), ActionExplanations.class);
         } catch (IOException e) {
             log.error("Action explanations file not found {}", templatesPath);
             throw new RuntimeException(e);